move `Notedeck` to `notedeck` crate
