+++ .gitignore
@@ -17,3 +17,4 @@ scripts/macos_build_secrets.sh
 *.txt
 /tags
 *.mdb
+.idea/
