+++ crates/notedeck/src/lib.rs
@@ -9,19 +9,20 @@ mod imgcache;
 mod muted;
 pub mod note;
 mod notecache;
+mod persist;
 mod result;
 pub mod storage;
 mod style;
 pub mod theme;
-mod theme_handler;
 mod time;
 mod timecache;
+mod timed_serializer;
 pub mod ui;
 mod unknowns;
 mod user_account;
 
 pub use accounts::{AccountData, Accounts, AccountsAction, AddAccountAction, SwitchAccountAction};
-pub use app::App;
+pub use app::{App, Notedeck};
 pub use args::Args;
 pub use context::AppContext;
 pub use error::{Error, FilterError};
@@ -31,13 +32,13 @@ pub use imgcache::ImageCache;
 pub use muted::{MuteFun, Muted};
 pub use note::{NoteRef, RootIdError, RootNoteId, RootNoteIdBuf};
 pub use notecache::{CachedNote, NoteCache};
+pub use persist::*;
 pub use result::Result;
 pub use storage::{
     DataPath, DataPathType, Directory, FileKeyStorage, KeyStorageResponse, KeyStorageType,
 };
 pub use style::NotedeckTextStyle;
 pub use theme::ColorTheme;
-pub use theme_handler::ThemeHandler;
 pub use time::time_ago_since;
 pub use timecache::TimeCached;
 pub use unknowns::{get_unknown_note_ids, NoteRefsUnkIdAction, SingleUnkIdAction, UnknownIds};
