+++ crates/notedeck/src/persist/mod.rs
@@ -1,0 +1,7 @@
+mod app_size;
+mod theme_handler;
+mod zoom;
+
+pub use app_size::AppSizeHandler;
+pub use theme_handler::ThemeHandler;
+pub use zoom::ZoomHandler;
