+++ crates/notedeck_chrome/Cargo.toml
@@ -12,8 +12,6 @@ description = "The nostr browser"
 eframe = { workspace = true }
 egui_extras = { workspace = true }
 egui = { workspace = true }
-enostr = { workspace = true }
-nostrdb = { workspace = true }
 notedeck_columns = { workspace = true }
 notedeck = { workspace = true }
 puffin = { workspace = true, optional = true }
