+++ crates/notedeck_chrome/src/lib.rs
@@ -1,13 +1,6 @@
-pub mod app_size;
 pub mod fonts;
-pub mod persist_zoom;
 pub mod setup;
 pub mod theme;
-pub mod timed_serializer;
-
-mod app;
-
-pub use app::Notedeck;
 
 #[cfg(target_os = "android")]
 mod android;
